requirements.txt
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pdfplumber>=0.9.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
click>=8.1.0
pydantic>=2.0.0
typing-extensions>=4.5.0
streamlit>=1.28.0
plotly>=5.15.0
